Refactor creep inheritance hierarchy
