+++ src/modules/creepDriver.ts
@@ -1,21 +1,30 @@
-import { Maintainer } from '../roles/maintainer';
+import { EarlyMaintainer } from '../roles/earlyMaintainer';
+import { EarlyUpgrader } from '../roles/earlyUpgrader';
+import { EarlyDrone } from '../roles/earlyDrone';
+import { WaveCreep } from '../virtualCreeps/WaveCreep';
 import { Upgrader } from '../roles/upgrader';
-import { Worker } from '../roles/worker';
-import { WaveCreep } from './WaveCreep';
+import { Maintainer } from '../roles/maintainer';
 
 export default function driveCreep(creep: Creep) {
     let waveCreep: WaveCreep;
 
     switch (creep.memory.role) {
         case Role.WORKER:
-            waveCreep = new Worker(creep.id);
+            waveCreep = new EarlyDrone(creep.id);
             break;
         case Role.UPGRADER:
-            waveCreep = new Upgrader(creep.id);
+            if (Memory.rooms[creep.memory.room].phase === 1) {
+                waveCreep = new EarlyUpgrader(creep.id);
+            } else {
+                waveCreep = new Upgrader(creep.id);
+            }
             break;
         case Role.MAINTAINTER:
-            waveCreep = new Maintainer(creep.id);
-            break;
+            if (Memory.rooms[creep.memory.room].phase === 1) {
+                waveCreep = new EarlyMaintainer(creep.id);
+            } else {
+                waveCreep = new Maintainer(creep.id);
+            }
         default:
             waveCreep = new WaveCreep(creep.id);
     }
