+++ src/roles/earlyUpgrader.ts
@@ -1,0 +1,7 @@
+import { EarlyCreep } from '../virtualCreeps/earlyCreep';
+
+export class EarlyUpgrader extends EarlyCreep {
+    protected performDuties() {
+        this.runUpgradeJob();
+    }
+}
