+++ src/virtualCreeps/transporterCreep.ts
@@ -1,0 +1,3 @@
+import { WaveCreep } from './waveCreep';
+
+export class TransporterCreep extends WaveCreep {}
